+++ mne_bids/tests/test_write.py
@@ -42,7 +42,6 @@ from mne_bids import (write_raw_bids, read_raw_bids, BIDSPath,
                       get_anat_landmarks, write, anonymize_dataset,
                       get_entity_vals)
 from mne_bids.write import _get_fid_coords
-from mne_bids.dig import _write_dig_bids, _read_dig_bids
 from mne_bids.utils import (_stamp_to_dt, _get_anonymization_daysback,
                             get_anonymization_daysback, _write_json)
 from mne_bids.tsv_handler import _from_tsv, _to_tsv
@@ -1312,8 +1311,8 @@ def test_eegieeg(dir_name, fname, reader, _bids_validate, tmp_path):
                                                 coord_frame='head')
     raw.set_montage(eeg_montage)
     # electrodes are not written w/o landmarks
-    with pytest.warns(RuntimeWarning, match='Skipping EEG electrodes.tsv... '
-                                            'Setting montage not possible'):
+    with pytest.raises(RuntimeError, match="'head' coordinate frame must "
+                                           "contain nasion"):
         write_raw_bids(raw, bids_path, overwrite=True)
 
     electrodes_fpath = _find_matching_sidecar(bids_path,
@@ -1638,8 +1637,8 @@ def test_snirf(_bids_validate, tmp_path):
 
     raw = _read_raw_snirf(raw_fname, optode_frame="mri")
     raw.info['dig'].pop(1)
-    with pytest.warns(RuntimeWarning,
-                      match='Setting montage not possible'):
+    with pytest.raises(RuntimeError,
+                       match="'head' coordinate frame must contain nasion"):
         write_raw_bids(raw, bids_path, overwrite=True)
 
 
@@ -3611,43 +3610,4 @@ def test_anonymize_dataset_daysback(tmpdir):
         ],
         rng=np.random.default_rng(),
         show_progress_thresh=20
-    )
-
-
-def test_write_dig(tmpdir):
-    """Test whether the channel locations are written out properly."""
-    # Check progress bar output
-    bids_root = tmpdir / 'bids'
-    data_path = Path(testing.data_path())
-    raw_path = data_path / 'MEG' / 'sample' / 'sample_audvis_trunc_raw.fif'
-
-    # test coordinates in pixels
-    bids_path = _bids_path.copy().update(
-        root=bids_root, datatype='ieeg', space='Pixels')
-    os.makedirs(op.join(bids_root, 'sub-01', 'ses-01', bids_path.datatype),
-                exist_ok=True)
-    raw = _read_raw_fif(raw_path, verbose=False)
-    raw.pick_types(eeg=True)
-    raw.del_proj()
-    raw.set_channel_types({ch: 'ecog' for ch in raw.ch_names})
-
-    montage = raw.get_montage()
-    # fake transform to pixel coordinates
-    montage.apply_trans(mne.transforms.Transform('head', 'unknown'))
-    with pytest.warns(RuntimeWarning,
-                      match='assuming identity'):
-        _write_dig_bids(bids_path, raw, montage)
-    electrodes_path = bids_path.copy().update(
-        task=None, run=None, suffix='electrodes', extension='.tsv')
-    coordsystem_path = bids_path.copy().update(
-        task=None, run=None, suffix='coordsystem', extension='.json')
-    with pytest.warns(RuntimeWarning,
-                      match='not an MNE-Python coordinate frame'):
-        _read_dig_bids(electrodes_path, coordsystem_path,
-                       bids_path.datatype, raw)
-    montage2 = raw.get_montage()
-    assert montage2.get_positions()['coord_frame'] == 'unknown'
-    assert_array_almost_equal(
-        np.array(list(montage.get_positions()['ch_pos'].values())),
-        np.array(list(montage2.get_positions()['ch_pos'].values()))
     )